+++ fast-ftrl-proximal/src/main/java/io/scaledml/FtrlProximalRunner.java
@@ -4,9 +4,9 @@ package io.scaledml;
 import com.lmax.disruptor.dsl.Disruptor;
 import io.scaledml.io.LineBytesBuffer;
 
-import java.nio.ByteBuffer;
-
-public class ParallelRun {
+public class FtrlProximalRunner {
     Disruptor<LineBytesBuffer> inputDisruptor;
     Disruptor<FtrlProximalState.Increment> incrementDisruptor;
+
+    
 }
