+++ fast-ftrl-proximal/src/main/java/io/scaledml/InputFormat.java
@@ -1,0 +1,10 @@
+package io.scaledml;
+
+import io.scaledml.io.LineBytesBuffer;
+
+/**
+ * Created by aonuchin on 24.03.15.
+ */
+public interface InputFormat {
+    SparseItem parse(LineBytesBuffer line);
+}
